node: regex support for action variable mapping
The action mapping string can now contain chained (almost) arbitrary
regular expressions to manipulate the input. The regular expressions are
applied in the order given and before a modifier is applied, e.g.,
converting everything to an uppercase string (or keeping the string
as-is if desired).

An example could be for an action:
(move ?r - robot ?from ?to - location)
to convert to a string using:
 move -> ppgoto{place=?(to|/c-(.+)/m-$1/)S}
This would use the destination as the argument to the ppgoto skill.
Additionally, it would convert a leading "c-" to a leading "m-" and
after substitution replace the string by its uppercase version. For
example, an action such as (move r-1 start c-bs-i) to move robot r-1
from the location start to the location c-bs-i would be converted to
'ppgoto{place="M-BS-I"}, note the replacement, modifier, and addition of
quotation marks (through the S modifier).

For simplicity on the parsing side, regular expressions may contain
neither slash (/) nor pipe (|) symbols.