node: support always suceeding actions
Support actions which are considered to always succeed. This is useful,
for example, for actions which simply set a fluent or predicate. For
example, a mark-visited "action" might be used to mark a certain place
as having been visited, but there is no corresponding skill to execute.
Note that the exact identifier variable filtering applies as for mapped
actions.