node: fix effect handling (and workaround ROSPlan constants)
The effects were not properly handled, some values were taken from the
wrong place.

A particular problem are constants. They are not explicitly represented
currently by ROSPlan. Hence, just the absence of a value in the bounded
parameters can be taken as a hint that a value is actually a constant.
The same problem should exist in the provided RPActionInterface.