kb_updater_navgraph: assert a distance function based on navgraph
This node can be used to write pairwise distances (along the paths in
the navgraph, not necessarily straight line distance) to the ROSPlan
knowledge base. It monitors the machine info topic to see if more
information is available and the navgraph topic to detect updated
navgraphs. It makes as few modifications to the KB as possible (you
should only ever see it updated once on starting the game, or multiple
times during exploration).