action-dispatcher: support dashes in action names
Ditch the simple ROS API for retrieving action configuration. That
allows to get rid of the translation of dashes to underscores. That is
an arbitrary assumption and I don't really see why ROS makes it.

With this we can have a more natural configuration and a one-to-one
mapping for PDDL operators that contain one or more dashes in the name.